Windows pass: Now buildable with VC++ Express 2010 No longer requires boost_thread, using SDL instead but some boost_thread code lingers behind an #ifdef Changed all gl buffer calls/constants to be non-ARB (I think) which works better on especially crappy PC hardware (tested on 2-yr-old Samsung netbook) etc.
Note: Is probably now broken on *NIX

git-svn-id: https://crag.googlecode.com/svn/trunk@5 03cb304c-929a-b8d2-5531-aad0e9782889

+++ src/crag/gfx/Font.h
@@ -23,10 +23,10 @@ namespace gfx
 	public:
 		Font(char const * filename, float scale = 1.f);
 
-		void Print(char const * text, Vector2i const & position) const;
+		void Print(char const * text, Vector2f const & position) const;
 
 	private:
-		void GenerateVerts(char const * text, Vector2i const & position) const;
+		void GenerateVerts(char const * text, Vector2f const & position) const;
 		void RenderVerts() const;
 		
 		void PrintChar(char c, Vector2f & position) const;
